player-prop-gnn/requirements.txt
# Core
numpy==1.26.0
pandas==2.1.0
scipy==1.11.0
scikit-learn==1.3.0

# Data
statsbombpy==1.11.0
requests==2.31.0
beautifulsoup4==4.12.2

# Bayesian
pymc==5.9.0
arviz==0.16.1

# Deep Learning - PyTorch
torch==2.1.0

# Deep Learning - PyTorch Geometric (GNN)
torch-geometric==2.4.0
torch-scatter==2.1.1
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
networkx==3.2

# Dashboard
streamlit==1.28.0

# Testing
pytest==7.4.3

# Utils
python-dotenv==1.0.0
tqdm==4.66.1
pydantic-settings==2.0.3
redis==5.0.0
psutil==5.9.6